Un massiv refactor to get it to work
