+++ frontend/src/views/ProjectsView.tsx
@@ -1,23 +1,49 @@
-import ProjectList from "../components/ProjectList.tsx";
 import "../styles/ProjectView.css";
-import "../styles/ProjectList.css";
 import CreateProjectDialog from "../components/CreateProjectDialog.tsx";
 import { Token } from "../contexts/Auth";
 import JoinProjectDialog from "../components/JoinProjectDialog.tsx";
+import { useState, useEffect } from "react";
+import "../styles/ProjectList.css";
+import call, { Method } from "../utils/api.ts";
+import Project from "../components/Project.tsx";
 
 interface ProjectsViewProps {
     token: Token;
 }
 
+interface Project {
+    projectid: string;
+}
+
 export default function ProjectsView({ token }: Readonly<ProjectsViewProps>) {
+    const [projects, setProjects] = useState<Project[]>([]);
+
+    useEffect(() => {
+        call("/project", Method.Get, token)
+            .then((response) => response.json())
+            .then((data: Project[]) => {
+                setProjects(data);
+            })
+            .catch((err) => console.log(err));
+    }, []);
+
     return (
         <div className="container">
-            <div className="project-list-div">
-                <ProjectList token={token} />
-            </div>
             <div className="project-controls">
                 <CreateProjectDialog token={token} />
                 <JoinProjectDialog token={token} />
+            </div>
+            <div className="project-list-div">
+                <h2>Existing Projects</h2>
+                <ul style={{ listStyle: "none" }}>
+                    {projects.map((project) => (
+                        <Project
+                            key={project.projectid}
+                            token={token}
+                            id={project.projectid}
+                        />
+                    ))}
+                </ul>
             </div>
         </div>
     );
